Enhancement/refactor products

+++ lib/presentation/base_data/edit_product/widgets/edit_threshold.dart
@@ -14,7 +14,7 @@ class EditThreshold extends HookConsumerWidget {
           initialValue: _product.value.threshold.toString(),
           onChanged: (value) {
             _product.value =
-                _product.value.copyWith(threshold: double.parse(value));
+                _product.value.copyWith(threshold: int.parse(value));
           },
           keyboardType: TextInputType.number,
           decoration: const InputDecoration(
